Call Link Page Updated
